refactor: Replace type switches with interface methods and extract utilities
Add CelExpressions() and PopulateProtobuf() to the Action interface,
eliminating type switches in the SDK controller. Replace the
validateActions switch with a validator function table for
extensibility. Extract AST and proto descriptor helpers from manager.go
into cel_utils.go and proto_utils.go.


rh-pre-commit.version: 2.3.2
rh-pre-commit.check-secrets: ENABLED
